Revert "feat: 인생지도 공유 url 화면 추가"
