+++ src/app/home/layout.tsx
@@ -9,9 +9,7 @@ const HomeLayout = ({ children }: PropsWithChildren) => {
       <div className="absolute right-0">
         <Image src={PurpleBlurImage} width={390} height={125} alt="purple_blur" />
       </div>
-      <div className="w-full h-[100dvh]">
-        <div className="w-full h-[100dvh] flex flex-col items-center justify-between pb-xs">{children}</div>
-      </div>
+      <div className="w-full h-[100dvh]">{children}</div>
     </main>
   );
 };
