+++ src/app/home/page.tsx
@@ -1,0 +1,7 @@
+import { LifeMap } from '@/features/home/components';
+
+const HomePage = () => {
+  return <LifeMap />;
+};
+
+export default HomePage;
