+++ src/app/home/startBg.tsx
