+++ src/features/goal/components/detail/DetailHeader.tsx
@@ -4,7 +4,6 @@ import { useOverlay } from '@toss/use-overlay';
 
 import CloseIcon from '@/assets/icons/goal/close-icon.svg';
 import DeleteIcon from '@/assets/icons/goal/delete-icon.svg';
-import { useGetMemberData } from '@/hooks/reactQuery/auth';
 
 import { DeleteGoalButtomSheet } from './DeleteGoalButtomSheet';
 
@@ -14,11 +13,10 @@ interface DetailHeaderProps {
 
 export const DetailHeader = ({ goalId }: DetailHeaderProps) => {
   const overlay = useOverlay();
-  const { data } = useGetMemberData();
 
   return (
     <>
-      <Link href={{ pathname: `/home/${data?.username}` }}>
+      <Link href={{ pathname: '/home' }}>
         <CloseIcon />
       </Link>
