+++ src/features/goal/components/detail/SavedFooterButton.tsx
@@ -2,13 +2,10 @@ import React from 'react';
 import Link from 'next/link';
 
 import { Button } from '@/components';
-import { useGetMemberData } from '@/hooks/reactQuery/auth';
 
 export const SavedFooterButton = () => {
-  const { data: memberData } = useGetMemberData();
-
   return (
-    <Link href={{ pathname: `/home/${memberData?.username}` }}>
+    <Link href={{ pathname: '/home' }}>
       <Button>홈으로 가기</Button>
     </Link>
   );
