+++ src/features/my/components/MyPageLayout.tsx
@@ -14,7 +14,7 @@ export const MyPageLayout = () => {
 
   return (
     <div className="pt-5xs px-xs h-full flex flex-col">
-      <MypageHeader username={memberData?.username} />
+      <MypageHeader />
       {memberData && (
         <>
           <UserProfile
