+++ src/hooks/reactQuery/goal/useGetGoals.ts
@@ -1,8 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { useAtomValue } from 'jotai';
 
 import { api } from '@/apis';
-import { isLoginAtom } from '@/features/auth/atom';
 
 export type GoalProps = {
   id: number;
@@ -18,11 +16,8 @@ export type GoalResponse = {
 };
 
 export const useGetGoals = () => {
-  const isLogin = useAtomValue(isLoginAtom);
-
   return useQuery<GoalResponse>({
     queryKey: ['goals'],
     queryFn: () => api.get<GoalResponse>('/life-map'),
-    enabled: isLogin,
   });
 };
